requirements.txt
qiskit>=1.0
qiskit-algorithms>=0.3.0
qiskit-optimization>=0.6.0
qiskit-aer>=0.15
qiskit-ionq>=0.5
qcentroid-runtime>=0.1.4
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
ViennaRNA>=2.6
pytest>=7.0
jupyter>=1.0
